LayoutTests/imported/w3c/web-platform-tests/css/css-tables/visibility-collapse-colspan-002-expected.txt

Visibility collapse

Spec
Setting a spanning column to visibility:collapse changes table width but not height. The first column has been collapsed and column width is the widest cell in the column.


PASS col visibility:collapse doesn't change table height
FAIL col visibility:collapse changes table width assert_equals: col visibility:collapse changes table width expected 50 but got 60
